requirements.txt
fastapi
fastapi[standard]
uvicorn
openai
google-generativeai
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
selenium
webdriver-manager
